+++ QuiltServer/ChangeLog.md
@@ -18,6 +18,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/)
   - Arrows can now be used as fuel with same fuel value as sticks
   - Crafting recipes with chest input now accept any chest
   - Decorative items from This Rocks! mod now drop themselves and can be placed by players
+  - Modded fish can now be used in any recipe with input of raw fish
+  - Modded cooked fish can now be used in any recipe with input of cooked fish
 
 ### Changed